backend/requirements.txt
django==1.9.7
django-environ==0.4.0
django-filter==0.13.0
djangorestframework==3.3.3
Markdown==2.6.6
